+++ lib/src/main/java/com/smileidentity/viewmodel/BiometricKycViewModel.kt
@@ -80,7 +80,8 @@ class BiometricKycViewModel(
             val didMoveToSubmitted: Boolean
             if (useStrictMode) {
                 didMoveToSubmitted =
-                    handleOfflineJobFailure(jobId, e) && handleOfflineJobFailure(userId, e)
+                    handleOfflineJobFailure(jobId, e) &&
+                    handleOfflineJobFailure(userId, e)
             } else {
                 didMoveToSubmitted = handleOfflineJobFailure(jobId, e)
             }